refactor(rest-api): move file_auth actors in correct location
