+++ rest-api/src/actors.rs
@@ -1,6 +1,2 @@
-pub mod git_actor;
-pub mod github_release_actor;
 pub mod nonce_cache_actor;
 pub mod nonce_cleanup_actor;
-pub mod signature_collector;
-pub mod signers_initialiser;
