+++ rest-api/src/file_auth/actors.rs
@@ -1,1 +1,5 @@
 pub mod forge_signers_validator;
+pub mod git_actor;
+pub mod github_release_actor;
+pub mod signature_collector;
+pub mod signers_initialiser;
